Display an alert if no keycloak server is running.
This is because I'm a bonehead and always forget to start keycloak when
I'm working on this project. Then I spend 10 minutes trying to figure
out how/why it's broken before realizing that I just never started the
damn keycloak server. So now, I'll know right off the bat. Yay!