Translate service portal, pod, and host IPs to external service names
Depending on how we found the internal service host we may have a
service's portal IP, a pod's cluster IP, or a pod's host IP that we
need to translate to an external service host. This basically combines
the changes in the last commit with the previous solution to handle
all three of these cases.